requirements.txt
click==8.0.3
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
prometheus-client==0.11.0
RPi.GPIO==0.7.1
Werkzeug==2.0.1